backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.31.0
pytest==8.0.2
PyJWT==2.8.0
mysql-connector-python==8.3.0
google-genai==0.8.0
gunicorn==23.0.0
